Fix some harmless compiler warnings

+++ app/streaming/input/keyboard.cpp
@@ -114,7 +114,7 @@ void SdlInputHandler::performSpecialKeyCombo(KeyCombo combo)
             }
 
             // Send this text to the PC
-            LiSendUtf8TextEvent(text, strlen(text));
+            LiSendUtf8TextEvent(text, (unsigned int)strlen(text));
 
             // SDL_GetClipboardText() allocates, so we must free
             SDL_free((void*)text);
